Support automatically converting between camel and snake case
This allows the Javascript/JSON convention of camelCased keys to be
supported alongside the Python convention of snake_cased variable
names. Without any extra thought by the user.

Humps seems a perfectly good library for converting, it is small and
has no additional dependencies.

Camelizing the full JSON Schema should be fine as JSONSchema is
defined with camelCased keys.