Support stdlib dataclasses
As it is possible to simply wrap a stdlib dataclass with the pydantic
dataclass to turn it into a pydantic dataclass the API can be changed
to accept stdlib dataclasses as well.

As stdlib dataclasses are a clearer usecase the documentation has been
changed to emphasise them as the prefered usage.