Add a `quart schema` command
This is available from the command line and will print the QUART_APP's
schema if available either to stdout or to a file with the
`-o`/`--output` option.

This results in the app storing a reference to the extension (rather
than the extension storing a reference to the app as before).