Add an info argument and model for OpenAPI info
This is a backwards incompatible change as the title and version
arguments have been removed and replaced with the info argument.

This starts an adoption of Pydantic models for OpenAPI objects. This
should help the user as validation errors will inform them of issues
with the provided data (before a OpenAPI parser fails trying to parse
it).

It isn't clear why the treat warnings as error aspect in the sphinx
build needs to be turned off.