Consistently apply casing conversion
Rather than alter the raw values as provided/used by Quart (via the
json, querystring or response values) the decorators and functions
convert as required. This should make the extension easier to use - as
adding custom JSONProviders with the previous setup was a pain.

This is backwards incompatible if the previous JSONProvider was used.

Note this may cause issues as the pydantic JSON encoding is no longer
used by default, which could cause issues when encoding - however it
should now be clear to a user how to fix this easily (no weird
inheritance or mixins).

I've also ensured that WebSockets and the test client follow the
casing conversion, so as to ensure that everything is consistent.