Fix failing validation tests for responses.
Fixes the problem with response validators that return either a list[T]
or a RootModel[T]. We now build a separate model_validator method at
the top of the ```validate_response``` decorator method that will use
the ```model_validate``` method if the model class is a pydantic
```RootModel``` or ```BaseModel``` instance. Otherwise, create a pydantic
```TypeAdapter``` that will attempt to build the model class from the
provided python data object.