Isolate the pydantic code to the conversion module
This should allow for further work to allow the extension to depend on
various validation libraries.

This in addition doesn't convert stdlib-dataclasses to pydantic
dataclasses which was done to take advantage of pydantic's field
validation, but I think it was a mistake as it changes the users
expectation (stdlib-dataclasses don't validated) and ties this
extension to pydantic.